requirements.txt
asgiref==3.10.0
crispy-tailwind==1.0.3
Django==5.2.8
django-ckeditor==6.7.3
django-crispy-forms==2.5
django-js-asset==3.1.2
numpy==2.3.4
pandas==2.3.3
pillow==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.10
dj-database-url==2.3.0
